requirements.txt
Django==2.2.28
djangorestframework==3.12.4
gunicorn==20.1.0
Pillow==7.2.0
pytz==2022.7
sqlparse==0.4.3
whitenoise==5.3.0
